day: 13/10/25

+++ 1300/EvenOdds.cpp
@@ -2,10 +2,10 @@
 using namespace std;
 
 int main(){
-    int n, k;
+    long long n, k;
     cin >> n >> k;
     
-    int oddCount = (n + 1) / 2;
+    long long oddCount = (n + 1) / 2;
     if(k <= oddCount){
         cout << 2 * k - 1 << endl;
     }
